+++ lib/redux/reducers/account_reducer.dart
@@ -7,7 +7,8 @@ final accountIdReducer = TypedReducer<int, ChangeAccount>(_accountIdReducer);
 final accountInfoIdReducer = TypedReducer<int, AccountInfoId>(_accountInfoIdReducer);
 final accountInfoNameReducer = TypedReducer<TextEditingController, AccountInfoName>(_accountInfoNameReducer);
 final accountInfoAcronymReducer = TypedReducer<TextEditingController, AccountInfoAcronym>(_accountInfoAcronymReducer);
-final accountInfoCurrencyRenducer = TypedReducer<String, AccountInfoCountryIso>(_accountInfoCurrencyReducer);
+final accountInfoCurrencyReducer = TypedReducer<String, AccountInfoCountryIso>(_accountInfoCurrencyReducer);
+final accountInfoCurrencyTextReducer = TypedReducer<TextEditingController, AccountInfoCurrencyText>(_accountInfoCurrencyTextReducer);
 
 int _accountIdReducer(int accountId, ChangeAccount action) {
   return action.accountId;
@@ -25,6 +26,10 @@ TextEditingController _accountInfoAcronymReducer(TextEditingController accountIn
   return action.accountInfoAcronym;
 }
 
-String _accountInfoCurrencyReducer(String accountInfoCurrency, AccountInfoCountryIso action) {
+String _accountInfoCurrencyReducer(String accountInfoCountryIso, AccountInfoCountryIso action) {
   return action.accountInfoCountryIso;
+}
+
+TextEditingController _accountInfoCurrencyTextReducer(TextEditingController accountInfoCurrencyText, AccountInfoCurrencyText action) {
+  return action.accountInfoCurrencyText;
 }